requirements.txt
ipython==8.12.3
lexicalrichness==0.5.1
librosa==0.11.0
matplotlib==3.10.7
noisereduce==3.0.3
numpy==1.26.4
opensmile==2.6.0
pandas==2.2.3
praatio==6.2.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyannote_audio==3.4.0
pyloudnorm==0.2.0
PyYAML==6.0.3
scipy==1.13.1
scikit-learn==1.5.2
seaborn==0.13.2
soundfile==0.13.1
spacy==3.8.11
speechbrain==1.0.3
tgt==1.5
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
transformers==4.57.1
webrtcvad_wheels==2.0.14
xgboost==3.1.2